command baseclasse

+++ srcs/commands/0-Command.cpp
@@ -15,6 +15,6 @@ vector<string> Command::getCommand(string input_client){
 	stringstream ss(input_client);
 	while(getline(ss, tok, ' ')){ split.push_back(tok);	}
 	cout  << "verif" << endl;
-	for (int i = 0;  i < split.size(); i++){ cout << split[i] << endl; }
+	for (size_t i = 0;  i < split.size(); i++){ cout << split[i] << endl; }
 	return (split);
 }
